Description
An interactive flask app is built which takes up responses from the user and is preprocessed for NLP. Covid related data are framed into a question-answer format in a json file with a unique tagging for each class. User input is fed into a neural network and predicts the appropriate class. Based on the predicted class label, the model chooses a response and is displayed. App also has a gallery page and an option which is directly linked to WHO page.